Externalize `production/data`
