+++ production/skilljar-theme-v3.0/static.mjs
@@ -23,7 +23,7 @@ export const CONFIG = {
       el("path", {
         d: "M315.566 78.3906C312.897 78.3906 310.475 77.7656 308.301 76.5156C306.139 75.2656 304.44 73.5404 303.203 71.3398C301.979 69.1393 301.367 66.6914 301.367 63.9961C301.367 61.3008 301.979 58.8529 303.203 56.6523C304.44 54.4518 306.139 52.7266 308.301 51.4766C310.475 50.2266 312.897 49.6016 315.566 49.6016C317.741 49.6016 319.798 50.0443 321.738 50.9297C323.678 51.8021 325.286 53.026 326.562 54.6016C327.852 56.1641 328.626 57.9089 328.887 59.8359H322.285C321.908 58.5208 321.094 57.4466 319.844 56.6133C318.607 55.7799 317.181 55.3633 315.566 55.3633C313.262 55.3633 311.328 56.1901 309.766 57.8438C308.216 59.4844 307.441 61.5352 307.441 63.9961C307.441 66.457 308.216 68.5143 309.766 70.168C311.328 71.8216 313.262 72.6484 315.566 72.6484C317.194 72.6484 318.633 72.1927 319.883 71.2812C321.146 70.3698 321.96 69.2109 322.324 67.8047H328.965C328.431 70.9167 326.914 73.4622 324.414 75.4414C321.927 77.4076 318.978 78.3906 315.566 78.3906ZM339.845 57.9609C341.265 57.9609 342.599 58.2214 343.849 58.7422C345.112 59.263 346.193 59.9792 347.091 60.8906C348.003 61.7891 348.719 62.8763 349.24 64.1523C349.774 65.4154 350.041 66.763 350.041 68.1953C350.041 70.1094 349.591 71.8477 348.693 73.4102C347.808 74.9727 346.584 76.1966 345.021 77.082C343.472 77.9544 341.746 78.3906 339.845 78.3906C336.942 78.3906 334.513 77.4206 332.56 75.4805C330.607 73.5404 329.63 71.112 329.63 68.1953C329.63 65.2786 330.607 62.8438 332.56 60.8906C334.513 58.9375 336.942 57.9609 339.845 57.9609ZM339.845 63.3516C338.504 63.3516 337.384 63.8138 336.486 64.7383C335.601 65.6497 335.158 66.8021 335.158 68.1953C335.158 69.5495 335.601 70.6888 336.486 71.6133C337.384 72.5378 338.504 73 339.845 73C341.16 73 342.267 72.5378 343.166 71.6133C344.064 70.6888 344.513 69.5495 344.513 68.1953C344.513 66.8021 344.064 65.6497 343.166 64.7383C342.28 63.8138 341.173 63.3516 339.845 63.3516ZM365.003 58.3516H370.648V78H365.003V75.4023C363.766 77.3945 361.995 78.3906 359.691 78.3906C357.178 78.3906 355.198 77.5964 353.753 76.0078C352.321 74.4193 351.605 72.2318 351.605 69.4453V58.3516H357.249V68.7227C357.249 70.0247 357.601 71.0664 358.304 71.8477C359.007 72.6159 359.951 73 361.136 73C362.295 73 363.226 72.6159 363.929 71.8477C364.645 71.0664 365.003 70.0247 365.003 68.7227V58.3516ZM383.735 57.9609C384.139 57.9609 384.77 58.0391 385.63 58.1953V63.3125C384.835 63.1302 384.152 63.0391 383.579 63.0391C382.134 63.0391 380.949 63.4883 380.024 64.3867C379.113 65.2721 378.657 66.5807 378.657 68.3125V78H373.012V58.3516H378.657V60.5977C379.868 58.8398 381.561 57.9609 383.735 57.9609ZM394.108 78.3906C391.517 78.3906 389.401 77.7721 387.76 76.5352C386.133 75.2852 385.221 73.5599 385.026 71.3594H390.436C390.488 72.0234 390.833 72.5573 391.471 72.9609C392.109 73.3646 392.897 73.5664 393.834 73.5664C394.668 73.5664 395.312 73.4362 395.768 73.1758C396.237 72.9154 396.471 72.5508 396.471 72.082C396.471 71.6523 396.269 71.3008 395.866 71.0273C395.462 70.7539 394.935 70.5521 394.284 70.4219C393.633 70.2786 392.91 70.1419 392.116 70.0117C391.334 69.8815 390.553 69.6992 389.772 69.4648C388.991 69.2305 388.274 68.931 387.623 68.5664C386.972 68.2018 386.445 67.668 386.041 66.9648C385.638 66.2487 385.436 65.3958 385.436 64.4062C385.436 62.5443 386.139 61.0078 387.545 59.7969C388.965 58.5729 390.859 57.9609 393.229 57.9609C395.468 57.9609 397.389 58.5339 398.991 59.6797C400.592 60.8255 401.51 62.4401 401.745 64.5234H396.061C396.022 64.0026 395.735 63.5599 395.202 63.1953C394.668 62.8307 393.971 62.6484 393.112 62.6484C392.435 62.6484 391.888 62.7656 391.471 63C391.067 63.2214 390.866 63.5534 390.866 63.9961C390.866 64.3997 391.067 64.7253 391.471 64.9727C391.888 65.207 392.422 65.3828 393.073 65.5C393.724 65.6172 394.44 65.7344 395.221 65.8516C396.002 65.9688 396.784 66.151 397.565 66.3984C398.359 66.6458 399.082 66.9779 399.733 67.3945C400.384 67.7982 400.911 68.3971 401.315 69.1914C401.732 69.9857 401.94 70.9362 401.94 72.043C401.94 73.8398 401.185 75.3503 399.674 76.5742C398.177 77.7852 396.321 78.3906 394.108 78.3906ZM422.098 67.5117C422.098 68.2018 422.059 68.8724 421.98 69.5234H407.391C407.586 70.7995 408.139 71.8216 409.051 72.5898C409.962 73.3451 411.076 73.7227 412.391 73.7227C413.471 73.7227 414.35 73.5143 415.027 73.0977C415.717 72.668 416.141 72.1016 416.297 71.3984H421.824C421.499 73.5078 420.49 75.2005 418.797 76.4766C417.104 77.7526 415.021 78.3906 412.547 78.3906C410.646 78.3906 408.901 77.9414 407.312 77.043C405.737 76.1315 404.493 74.888 403.582 73.3125C402.671 71.724 402.215 69.9792 402.215 68.0781C402.215 65.2266 403.198 62.8307 405.164 60.8906C407.143 58.9375 409.565 57.9609 412.43 57.9609C414.174 57.9609 415.789 58.3776 417.273 59.2109C418.758 60.0443 419.93 61.1966 420.789 62.668C421.661 64.1393 422.098 65.7539 422.098 67.5117ZM407.664 65.9297H416.219C416.141 64.9531 415.711 64.1393 414.93 63.4883C414.148 62.8372 413.23 62.5117 412.176 62.5117C411.069 62.5117 410.125 62.7982 409.344 63.3711C408.562 63.944 408.003 64.7969 407.664 65.9297ZM431.455 78.3906C428.864 78.3906 426.748 77.7721 425.107 76.5352C423.479 75.2852 422.568 73.5599 422.373 71.3594H427.783C427.835 72.0234 428.18 72.5573 428.818 72.9609C429.456 73.3646 430.244 73.5664 431.181 73.5664C432.015 73.5664 432.659 73.4362 433.115 73.1758C433.584 72.9154 433.818 72.5508 433.818 72.082C433.818 71.6523 433.616 71.3008 433.213 71.0273C432.809 70.7539 432.282 70.5521 431.63 70.4219C430.979 70.2786 430.257 70.1419 429.463 70.0117C428.681 69.8815 427.9 69.6992 427.119 69.4648C426.338 69.2305 425.621 68.931 424.97 68.5664C424.319 68.2018 423.792 67.668 423.388 66.9648C422.985 66.2487 422.783 65.3958 422.783 64.4062C422.783 62.5443 423.486 61.0078 424.892 59.7969C426.311 58.5729 428.206 57.9609 430.576 57.9609C432.815 57.9609 434.736 58.5339 436.338 59.6797C437.939 60.8255 438.857 62.4401 439.091 64.5234H433.408C433.369 64.0026 433.082 63.5599 432.548 63.1953C432.015 62.8307 431.318 62.6484 430.459 62.6484C429.782 62.6484 429.235 62.7656 428.818 63C428.414 63.2214 428.213 63.5534 428.213 63.9961C428.213 64.3997 428.414 64.7253 428.818 64.9727C429.235 65.207 429.768 65.3828 430.42 65.5C431.071 65.6172 431.787 65.7344 432.568 65.8516C433.349 65.9688 434.13 66.151 434.912 66.3984C435.706 66.6458 436.429 66.9779 437.08 67.3945C437.731 67.7982 438.258 68.3971 438.662 69.1914C439.078 69.9857 439.287 70.9362 439.287 72.043C439.287 73.8398 438.532 75.3503 437.021 76.5742C435.524 77.7852 433.668 78.3906 431.455 78.3906Z",
       }),
-    ]
+    ],
   ),
   domains: {
     prod: { url: "courses.chainguard.dev", id: "3glgawqmzatte" },
@@ -175,672 +175,5 @@ export const bannerSVGs = {
   arrow: `<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 16 16" aria-hidden="true"><path d="M12.8 6.4V3.2H9.6v3.2h3.2ZM12.8 12.8V9.6H9.6v3.2h3.2ZM9.6 3.2V0H6.4v3.2h3.2ZM9.6 16v-3.2H6.4V16h3.2Z"/><path d="M9.6 9.6V6.4H6.4v3.2h3.2ZM16 9.6V6.4h-3.2v3.2H16ZM3.2 9.6V6.4H0v3.2h3.2Z"/></svg>`,
 };
 
-export const footerData = {
-  logo: {
-    href: "https://www.chainguard.dev/?utm_source=courses",
-    svg: `<svg height="20" viewBox="0 0 1051 170" fill="#14003D" xmlns="http://www.w3.org/2000/svg"><path fill-rule="evenodd" clip-rule="evenodd" d="M142.381 93.4236C143.741 88.7319 144.48 83.5906 144.48 77.9985C144.48 43.0218 115.557 3.65353 86.6691 3.65353C57.7816 3.65353 28.8578 43.0218 28.8578 77.9985C28.8578 85.6245 30.2328 92.4122 32.6826 98.3645L15.3182 97.3881C8.24572 96.9904 1.72988 102.033 2.83296 109.03C3.2608 111.744 4.07615 114.522 5.53233 116.921C0.616539 120.379 -1.39152 126.736 2.59162 131.718C6.59018 136.719 12.4611 141.327 20.5199 141.327C29.2607 141.327 34.5212 139.068 37.7159 136.231C37.9491 137.47 38.4301 138.69 39.1895 139.843C43.1126 145.799 49.61 152.346 59.2157 152.346C75.5561 152.346 77.3785 142.083 78.293 136.933C78.3645 136.53 78.4305 136.158 78.4974 135.825L89.4171 130.364L100.337 135.825C100.404 136.158 100.47 136.529 100.541 136.932L100.541 136.933C101.456 142.083 103.278 152.346 119.619 152.346C129.224 152.346 135.722 145.799 139.645 139.843C139.799 139.609 139.942 139.372 140.073 139.132C143.152 140.463 147.227 141.327 152.633 141.327C160.692 141.327 166.563 136.719 170.562 131.718C175.459 125.592 171.299 117.389 163.813 115.05L161.34 114.277C168.603 110.512 169.672 103.413 168.894 97.1917C168.015 90.1624 160.357 87.1238 153.672 89.4665L142.381 93.4236ZM89.4275 130.263C89.4211 130.263 89.4146 130.263 89.4081 130.263L89.4171 130.268L89.4275 130.263Z"></path><path d="M273.444 136.871C240.27 136.871 219.058 115.34 219.058 77.3815C219.058 41.4966 242.981 19.4871 274.56 19.4871C301.833 19.4871 321.45 35.117 324.321 61.5921H302.63C300.557 47.0787 290.988 37.1904 274.56 37.1904C254.784 37.1904 240.589 52.3418 240.589 77.3815C240.589 103.378 254.624 119.168 274.082 119.168C290.669 119.168 301.673 108.482 303.906 93.6494H324.799C322.088 119.487 303.428 136.871 273.444 136.871ZM340.007 134V22.3579H359.464V59.6783C359.464 61.9111 359.305 64.3035 358.826 67.3337C363.132 59.5188 370.469 53.6177 382.75 53.6177C399.974 53.6177 410.182 65.7389 410.182 84.8775V134H390.724V87.7483C390.724 76.2651 385.142 69.0881 376.051 69.0881C366.482 69.0881 359.464 77.541 359.464 88.3862V134H340.007ZM452.055 136.871C438.498 136.871 425.261 127.939 425.261 112.629C425.261 95.0848 438.817 89.0242 455.564 86.7914L466.728 85.356C473.107 84.5585 475.34 82.1662 475.34 78.3385C475.34 72.9158 470.875 68.1312 462.581 68.1312C453.49 68.1312 447.749 73.2348 446.951 81.3687H426.856C428.132 65.1009 441.369 53.6177 461.465 53.6177C485.228 53.6177 494.798 66.3768 494.798 88.2268V134H476.776V130.81C476.776 128.258 477.095 125.866 477.573 123.314C473.426 130.97 465.452 136.871 452.055 136.871ZM457.318 122.517C467.685 122.517 475.978 114.861 475.978 103.857V95.4037C474.224 96.9986 471.512 97.7961 466.728 98.753L460.348 100.029C451.257 101.783 445.037 104.814 445.037 111.831C445.037 118.689 450.46 122.517 457.318 122.517ZM513.824 134V56.1695H533.282V134H513.824ZM513.346 44.3674V23.9528H533.601V44.3674H513.346ZM553.385 134V56.1695H572.843V59.6783C572.843 61.9111 572.683 64.3035 572.205 67.3337C576.511 59.5188 583.847 53.6177 596.128 53.6177C613.353 53.6177 623.56 65.7389 623.56 84.8775V134H604.102V87.7483C604.102 76.2651 598.52 69.0881 589.429 69.0881C579.86 69.0881 572.843 77.541 572.843 88.3862V134H553.385ZM677.554 169.406C654.588 169.406 641.51 157.126 641.51 139.423H661.287C661.605 148.673 667.507 154.096 678.033 154.096C688.24 154.096 696.852 148.035 696.852 136.392V127.939C696.852 124.431 697.012 121.241 697.49 118.211C692.865 125.228 685.369 131.129 673.248 131.129C652.993 131.129 638.32 116.935 638.32 92.3735C638.32 68.1312 655.226 53.6177 672.291 53.6177C686.007 53.6177 692.865 59.1998 697.49 67.3337C697.012 64.6224 696.852 62.5491 696.852 59.6783V56.1695H716.31V131.608C716.31 157.126 700.521 169.406 677.554 169.406ZM678.192 115.659C689.675 115.659 697.969 105.93 697.969 92.3735C697.969 78.8169 689.675 69.0881 678.192 69.0881C666.39 69.0881 658.097 78.8169 658.097 92.3735C658.097 105.93 666.39 115.659 678.192 115.659ZM763.349 136.871C747.719 136.871 735.439 126.185 735.439 105.611V56.1695H755.056V102.581C755.056 113.426 759.841 121.241 769.25 121.241C779.139 121.241 786.156 113.904 786.156 102.421V56.1695H805.614V134H786.316V130.81C786.316 128.418 786.635 125.707 787.113 123.314C783.126 130.81 776.268 136.871 763.349 136.871ZM848.284 136.871C834.728 136.871 821.49 127.939 821.49 112.629C821.49 95.0848 835.047 89.0242 851.793 86.7914L862.957 85.356C869.337 84.5585 871.57 82.1662 871.57 78.3385C871.57 72.9158 867.104 68.1312 858.811 68.1312C849.72 68.1312 843.978 73.2348 843.181 81.3687H823.085C824.361 65.1009 837.599 53.6177 857.694 53.6177C881.458 53.6177 891.027 66.3768 891.027 88.2268V134H873.005V130.81C873.005 128.258 873.324 125.866 873.803 123.314C869.656 130.97 861.682 136.871 848.284 136.871ZM853.548 122.517C863.914 122.517 872.208 114.861 872.208 103.857V95.4037C870.453 96.9986 867.742 97.7961 862.957 98.753L856.578 100.029C847.487 101.783 841.267 104.814 841.267 111.831C841.267 118.689 846.69 122.517 853.548 122.517ZM910.054 134V56.1695H929.512V62.0706C929.512 64.9414 929.352 67.3337 928.874 70.2045C933.02 61.4327 940.038 53.6177 951.68 53.6177C953.275 53.6177 954.551 53.7772 955.827 54.0962V72.9158C954.551 72.5969 952.797 72.2779 950.245 72.2779C938.283 72.2779 929.512 78.3385 929.512 95.4037V134H910.054ZM997.069 136.871C976.495 136.871 961.503 122.836 961.503 95.4037C961.503 67.4932 979.206 53.6177 997.388 53.6177C1009.51 53.6177 1017 59.8378 1021.47 67.3337C1020.83 64.4629 1020.83 62.0706 1020.83 59.6783V22.3579H1040.29V134H1020.83V130.81C1020.83 128.099 1020.83 126.026 1021.47 123.314C1018.44 129.056 1011.42 136.871 997.069 136.871ZM1001.69 121.56C1013.34 121.56 1021.95 110.874 1021.95 95.2443C1021.95 79.2954 1013.34 69.0881 1001.69 69.0881C990.051 69.0881 981.279 79.7739 981.279 95.2443C981.279 111.353 990.051 121.56 1001.69 121.56Z"></path></svg>`,
-  },
-  socials: [
-    {
-      href: "https://www.linkedin.com/company/chainguard-dev/?utm_source=courses",
-      svg: `<svg xmlns="http://www.w3.org/2000/svg" width="25" height="24" viewBox="0 0 25 24" fill="none"><path d="M6.93866 0.54776H18.0816C19.6971 0.548746 21.2462 1.18917 22.3888 2.32844C23.5314 3.46772 24.174 5.01275 24.1758 6.62429V17.7458C24.174 19.3574 23.5314 20.9024 22.3888 22.0417C21.2462 23.1809 19.6971 23.8214 18.0816 23.8223H6.93866C5.32319 23.8214 3.77407 23.1809 2.6315 22.0417C1.48892 20.9024 0.846243 19.3574 0.844513 17.7458V6.62429C0.846243 5.01275 1.48892 3.46772 2.6315 2.32844C3.77407 1.18917 5.32319 0.548746 6.93866 0.54776Z" fill="#14003D" class="fill-[#14003D] dark:fill-white"></path><path d="M5.21951 9.74698H8.28983V19.8509H5.21951V9.74698Z" fill="white" class="fill-white dark:fill-[#14003D]"></path><path d="M10.2475 12.8723C10.2475 11.81 10.2475 10.5197 10.2195 9.58685H13.1966C13.267 10.0764 13.3075 10.5698 13.318 11.0643C13.7659 10.1985 14.7178 9.28799 16.9669 9.28799C19.4307 9.28799 20.2239 11.0569 20.2239 14.239V19.8426H17.1536V14.5918C17.1536 13.051 16.5376 11.9618 15.539 11.9618C13.8125 11.9618 13.3086 13.065 13.3086 15.3506V19.8426H10.2382V12.8723H10.2475Z" fill="white" class="fill-white dark:fill-[#14003D]"></path><path d="M6.38628 8.09361C6.74022 8.09325 7.08614 7.9882 7.38025 7.79177C7.67436 7.59534 7.90353 7.31635 8.03872 6.99003C8.17391 6.66372 8.20905 6.30473 8.13977 5.95848C8.07049 5.61222 7.89985 5.29425 7.64945 5.04471C7.39904 4.79517 7.08011 4.6253 6.73294 4.55655C6.38578 4.4878 6.02595 4.52326 5.69899 4.65846C5.37202 4.79366 5.09267 5.02255 4.89606 5.31615C4.69946 5.60976 4.59451 5.95491 4.59451 6.30799C4.59464 6.54261 4.6411 6.77489 4.73122 6.9916C4.82133 7.2083 4.95331 7.40518 5.11969 7.57099C5.28608 7.7368 5.48359 7.86831 5.70092 7.95798C5.91825 8.04765 6.15109 8.09374 6.38628 8.09361Z" fill="white" class="fill-white dark:fill-[#14003D]"></path></svg>`,
-    },
-    {
-      href: "https://www.youtube.com/@chainguard?utm_source=courses",
-      svg: `<svg xmlns="http://www.w3.org/2000/svg" width="32" height="22" viewBox="0 0 32 22" fill="none"><g clip-path="url(#clip0_560_16954)"><path class="dark:fill-white" d="M30.939 3.76117C30.765 3.10659 30.4251 2.51003 29.9533 2.03124C29.4815 1.55244 28.8944 1.20823 28.2508 1.03307C25.8804 0.387922 16.3768 0.387924 16.3768 0.387924C16.3768 0.387924 6.87064 0.387922 4.50282 1.03307C3.85919 1.20823 3.27207 1.55244 2.80028 2.03124C2.32848 2.51003 1.98856 3.10659 1.81459 3.76117C0.967769 8.67513 0.967769 13.7016 1.81459 18.6156C1.98856 19.2701 2.32848 19.8667 2.80028 20.3455C3.27207 20.8243 3.85919 21.1685 4.50282 21.3437C6.87325 21.9888 16.3768 21.9888 16.3768 21.9888C16.3768 21.9888 25.883 21.9888 28.2508 21.3437C28.8944 21.1685 29.4815 20.8243 29.9533 20.3455C30.4251 19.8667 30.765 19.2701 30.939 18.6156C31.7858 13.7016 31.7858 8.67513 30.939 3.76117ZM13.3341 15.8155V6.55948L21.232 11.1884L13.3341 15.8155Z" fill="#14003D"></path></g><defs><clipPath id="clip0_560_16954"><rect width="30.4" height="21.6" fill="white" transform="translate(0.820099 0.384995)"></rect></clipPath></defs></svg>`,
-    },
-    {
-      href: "https://x.com/chainguard_dev?utm_source=courses",
-      svg: `<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" style="/*! fill: black; */"><path d="M23.2 4.12826V20.2492C23.199 21.1972 22.8229 22.106 22.1544 22.776C21.4859 23.4459 20.5796 23.8223 19.6346 23.8223H3.56538C2.62042 23.8223 1.71412 23.4459 1.04559 22.776C0.377057 22.106 0.000985747 21.1972 1.9325e-06 20.2492V4.12826C-0.00048586 3.65823 0.0913754 3.19271 0.270334 2.75832C0.449292 2.32393 0.711839 1.92918 1.04296 1.59665C1.37409 1.26411 1.7673 1.00031 2.20011 0.820325C2.63292 0.64034 3.09685 0.547699 3.56538 0.547699L19.6346 0.547699C20.1032 0.547699 20.5671 0.64034 20.9999 0.820325C21.4327 1.00031 21.8259 1.26411 22.157 1.59665C22.4882 1.92918 22.7507 2.32393 22.9297 2.75832C23.1086 3.19271 23.2005 3.65823 23.2 4.12826Z" class="fill-[#14003D] dark:fill-white" style="fill: black;"></path><path d="M12.9469 11.4018L12.4476 10.6765L9.1811 5.92856H5.625L10.4562 13.0198L10.9545 13.7506L14.1635 18.4614H17.7939L12.9469 11.4018ZM11.49 13.1204L10.9925 12.3895L7.12836 6.71802H8.76629L11.9215 11.3003L12.4207 12.0265L16.2961 17.6626H14.5895L11.49 13.1204Z" class="fill-white dark:fill-[#14003D]" style="fill: #ffffff;"></path><path d="M10.7773 12.3836L11.2747 13.1145L10.7438 13.741L6.75345 18.4518H5.625L10.2455 13.0102L10.7773 12.3836Z" class="fill-white dark:fill-[#14003D]" style="fill: #ffffff;"></path><path d="M17.5451 5.92856L12.9051 11.4046L12.3743 12.0302L11.875 11.3041L12.4058 10.6785L16.4398 5.93043L17.5451 5.92856Z" class="fill-white dark:fill-[#14003D]" style="fill: #ffffff;"></path></svg>`,
-    },
-    {
-      href: "https://github.com/chainguard-dev?utm_source=courses",
-      svg: `<svg xmlns="http://www.w3.org/2000/svg" width="25" height="24" viewBox="0 0 25 24" fill="none"><g clip-path="url(#clip0_560_16956)"><path class="dark:fill-white" d="M20.9876 0.584808H4.11487C3.27582 0.584808 2.47114 0.918119 1.87784 1.51142C1.28454 2.10471 0.951233 2.9094 0.951233 3.74844L0.951233 20.6212C0.951233 21.4602 1.28454 22.2649 1.87784 22.8582C2.47114 23.4515 3.27582 23.7848 4.11487 23.7848H20.9876C21.8266 23.7848 22.6313 23.4515 23.2246 22.8582C23.8179 22.2649 24.1512 21.4602 24.1512 20.6212V3.74844C24.1512 2.9094 23.8179 2.10471 23.2246 1.51142C22.6313 0.918119 21.8266 0.584808 20.9876 0.584808ZM14.8251 19.1119C14.4534 19.183 14.3268 18.9589 14.3268 18.7731C14.3268 18.5411 14.3361 17.7752 14.3361 16.8222C14.3361 16.1539 14.1001 15.7175 13.847 15.504C15.4591 15.3287 17.1543 14.7263 17.1543 11.9897C17.165 11.2793 16.8983 10.5927 16.4109 10.0757C16.4834 9.89513 16.7286 9.17013 16.3384 8.1894C16.3384 8.1894 15.732 7.99826 14.3611 8.91967C13.1739 8.59979 11.9233 8.59979 10.7361 8.91967C9.34937 7.99694 8.74169 8.1894 8.74169 8.1894C8.34623 9.17276 8.59669 9.89776 8.67182 10.0757C8.18321 10.5911 7.91545 11.277 7.92573 11.9871C7.92573 14.7157 9.61828 15.3287 11.2212 15.5066C10.9583 15.7586 10.7954 16.0971 10.7625 16.4597C10.3472 16.6416 9.29664 16.9566 8.65337 15.8678C8.65337 15.8678 8.2711 15.1824 7.54346 15.1336C7.54346 15.1336 6.8356 15.1244 7.49205 15.5659C7.49205 15.5659 7.96923 15.7861 8.29878 16.6073C8.29878 16.6073 8.72455 17.9927 10.7414 17.5643C10.7414 18.1575 10.7493 18.6057 10.7493 18.7757C10.7493 18.9458 10.6175 19.1844 10.255 19.1171C7.37869 18.1786 5.30123 15.5093 5.30123 12.3614C5.30123 8.42667 8.54528 5.24326 12.5512 5.24326C16.5572 5.24326 19.8012 8.43326 19.8012 12.368C19.8012 15.504 17.7264 18.1759 14.8251 19.1119Z" fill="#14003D"></path></g><defs><clipPath id="clip0_560_16956"><rect width="23.2" height="23.2" fill="white" transform="translate(0.864532 0.584991)"></rect></clipPath></defs></svg>`,
-    },
-  ],
-  contact: {
-    href: "https://www.chainguard.dev/contact?utm_source=courses",
-    label: "Talk to an expert",
-  },
-  copyright: {
-    text: "©2025 Chainguard. All Rights Reserved.",
-    links: [
-      {
-        label: "Privacy",
-        href: "https://www.chainguard.dev/legal/privacy-notice?utm_source=courses",
-      },
-      {
-        label: "Terms",
-        href: "https://www.chainguard.dev/legal/terms-of-use?utm_source=courses",
-      },
-    ],
-  },
-  columns: [
-    {
-      className: "product-col",
-      groups: [
-        {
-          title: "Product",
-          links: [
-            {
-              label: "Chainguard Containers",
-              href: "https://www.chainguard.dev/containers",
-            },
-            {
-              label: "Chainguard Libraries",
-              href: "https://www.chainguard.dev/libraries",
-            },
-            { label: "Chainguard VMs", href: "https://www.chainguard.dev/vms" },
-            {
-              label: "Images Directory",
-              href: "https://images.chainguard.dev/",
-            },
-            {
-              label: "Integrations",
-              href: "https://www.chainguard.dev/scanners",
-            },
-            { label: "Pricing", href: "https://www.chainguard.dev/pricing" },
-          ],
-        },
-      ],
-    },
-    {
-      className: "solutions-col",
-      groups: [
-        {
-          title: "Solutions",
-          links: [
-            {
-              label: "FedRAMP",
-              href: "https://www.chainguard.dev/solutions/fedramp",
-            },
-            {
-              label: "PCI DSS",
-              href: "https://www.chainguard.dev/solutions/pci",
-            },
-            {
-              label: "CMMC 2.0",
-              href: "https://www.chainguard.dev/solutions/cmmc",
-            },
-            {
-              label: "Golden Images",
-              href: "https://www.chainguard.dev/solutions/golden-images",
-            },
-            {
-              label: "CVE Remediation",
-              href: "https://www.chainguard.dev/solutions/cve-remediation",
-            },
-            {
-              label: "Public Sector",
-              href: "https://www.chainguard.dev/solutions/public-sector",
-            },
-            {
-              label: "Startups",
-              href: "https://www.chainguard.dev/solutions/startups",
-            },
-          ],
-        },
-      ],
-    },
-    {
-      className: "resources-col",
-      groups: [
-        {
-          title: "Resources",
-          links: [
-            {
-              label: "Events & Webinars",
-              href: "https://www.chainguard.dev/events",
-            },
-            {
-              label: "Supply Chain Security 101",
-              href: "https://www.chainguard.dev/supply-chain-security-101",
-            },
-            {
-              label: "Chainguard Courses",
-              href: "https://courses.chainguard.dev/",
-            },
-            { label: "Documentation", href: "https://edu.chainguard.dev/" },
-            { label: "Trust Center", href: "https://edu.chainguard.dev/" },
-            {
-              label: "Chainguard Slack Community",
-              href: "https://communityinviter.com/apps/chainguardcommunity/invite",
-            },
-          ],
-        },
-        {
-          title: "Customers",
-          links: [
-            {
-              label: "Customer Stories",
-              href: "https://www.chainguard.dev/customers",
-            },
-            {
-              label: "Chainguard Reviews",
-              href: "https://www.chainguard.dev/chainguard-love",
-            },
-          ],
-        },
-      ],
-    },
-    {
-      className: "company-col",
-      groups: [
-        {
-          title: "Company",
-          links: [
-            {
-              label: "About Us",
-              href: "https://www.chainguard.dev/supply-chain-security-101",
-            },
-            { label: "Blog", href: "https://www.chainguard.dev/unchained" },
-            { label: "Partners", href: "https://www.chainguard.dev/partners" },
-            { label: "Newsroom", href: "https://www.chainguard.dev/newsroom" },
-            { label: "Careers", href: "https://www.chainguard.dev/careers" },
-            {
-              label: "Legal",
-              href: "https://www.chainguard.dev/legal/acceptable-use-policy",
-            },
-          ],
-        },
-      ],
-    },
-  ],
-};
-
-export const pathSections = {
-  home: [
-    {
-      internalSection: true,
-      eyebrow: "🔓 Internal Training",
-      title: "For Chainguardians",
-      description:
-        "Because you are logged in with a Chainguard email address, you can access internal training materials and resources to enhance your skills and knowledge about Chainguard's products and services.",
-      links: [
-        {
-          isPath: false,
-          isCourse: true,
-          hasBadge: false,
-          title: "Build Your First Chainguard Container",
-          slug: "build-your-first-chainguard-container",
-          description:
-            "This course is designed to teach new Chainguard engineers how to build container images.",
-          icon: "chainguard",
-        },
-        {
-          adminOnly: true,
-          isPath: true,
-          isCourse: false,
-          hasBadge: true,
-          icon: "burger",
-          title: "Painless Vulnerability Management",
-          slug: "path/painless-vulnerability-management",
-          description:
-            "Learn how to manage vulnerabilities effectively using Chainguard's tools and best practices. Currently, this is an internal learning path as we haven't rolled it out publicly yet.",
-        },
-      ],
-      classNames: ["internal"],
-    },
-    {
-      eyebrow: "Chainguard Fundamentals",
-      title: "Get Started with Containers",
-      description:
-        "Begin your Chainguard journey with essentials. Learn how containers work, how to use Chainguard securely, and gain confidence through guided onboarding.",
-      links: [
-        {
-          isPath: false,
-          isCourse: true,
-          hasBadge: false,
-          title: "Chainguard Containers Crash Course",
-          slug: "linkys-crash-course-on-chainguard-images",
-          description:
-            "A fast, focused overview to get you started—from setup and registry basics to vulnerability management and support.",
-          icon: "lightning",
-        },
-        {
-          isPath: true,
-          isCourse: false,
-          hasBadge: false,
-          title: "Chainguard Containers Onboarding Guide",
-          slug: "path/chainguard-containers-onboarding-guide",
-          description:
-            "A full 14-course path taking you from container image basics through migration, debugging, and registry mirroring.",
-          icon: "burger",
-        },
-      ],
-    },
-    {
-      eyebrow: "Tools & Customization",
-      title: "Make Chainguard Your Own",
-      description:
-        "Take control with customization tools. Learn to tailor Chainguard containers to your workflows and modernize Dockerfiles into secure, minimal builds.",
-      links: [
-        {
-          isPath: false,
-          isCourse: true,
-          hasBadge: false,
-          title: "Getting Started with Custom Assembly",
-          slug: "getting-started-with-chainguards-custom-assembly",
-          description:
-            "Quickly and securely customize your Chainguard Images—no Dockerfile required.",
-          icon: "star",
-        },
-        {
-          isPath: false,
-          isCourse: true,
-          hasBadge: false,
-          title: "Getting Started with Dockerfile Converter",
-          slug: "getting-started-with-chainguards-dockerfile-converter",
-          description:
-            "Convert Dockerfiles to secure, minimal Chainguard Containers with the Dockerfile Converter CLI.",
-          icon: "star",
-        },
-      ],
-    },
-    {
-      eyebrow: "Security & Compliance",
-      title: "Stay Ahead of Risks",
-      description:
-        "Level up your security practices. Understand vulnerabilities, compliance frameworks like SLSA, and secure emerging areas like AI/ML.",
-      links: [
-        {
-          isPath: true,
-          isCourse: false,
-          hasBadge: true,
-          icon: "burger",
-          title: "Painless Vulnerability Management",
-          slug: "path/painless-vulnerability-management",
-          description:
-            "Learn how to manage vulnerabilities effectively using Chainguard's tools and best practices.",
-        },
-        {
-          isPath: false,
-          isCourse: true,
-          hasBadge: false,
-          title: "Get Spicy with SLSA",
-          slug: "get-spicy-with-slsa-securing-your-supply-chain-one-level-at-a-time",
-          description:
-            "Understand SLSA, how Chainguard helps you meet it, and why it matters now more than ever.",
-          icon: "shield",
-        },
-        {
-          isPath: false,
-          isCourse: true,
-          hasBadge: true,
-          title: "Securing the AI/ML Supply Chain",
-          slug: "securing-ai",
-          description:
-            "Unpack threats, tools, and standards shaping MLSecOps—protect models, datasets, and AI/ML pipelines.",
-          icon: "frames",
-        },
-      ],
-    },
-
-    {
-      internalSection: true,
-      eyebrow: "Deep-Dive Paths",
-      title: "Master Containers",
-      description:
-        "Go beyond the basics with this in-depth learning path. Gain expertise in managing Chainguard Containers across the full software supply chain.",
-      links: [
-        {
-          isPath: true,
-          isCourse: false,
-          hasBadge: false,
-          title: "Complete Guide to Chainguard Containers",
-          slug: "path/linkys-guide-to-chainguard-images",
-          description:
-            "This learning path is being sunset. An 8-course path covering implementation, management, and best practices for Chainguard Containers.",
-          icon: "classBurger",
-        },
-      ],
-    },
-  ],
-  "painless-vulnerability-management": [
-    {
-      eyebrow: "",
-      title: "Painless Vulnerability Management",
-      description: "",
-      links: [
-        {
-          slug: "software-vulnerabilities-what-are-they",
-          isPath: false,
-          isCourse: true,
-          icon: "bookmark",
-          hasBadge: false,
-          title: "Software Vulnerabilities: What Are They?",
-          description:
-            "Understand the basics of software vulnerabilities, including common types and their implications.",
-        },
-        {
-          slug: "how-to-manage-cves",
-          isPath: false,
-          isCourse: true,
-          icon: "bookmark",
-          hasBadge: false,
-          title: "How to Manage CVEs",
-          description:
-            "Learn effective strategies for managing Common Vulnerabilities and Exposures (CVEs) in your software.",
-        },
-        {
-          slug: "more-secure-base-images",
-          isPath: false,
-          isCourse: true,
-          icon: "bookmark",
-          hasBadge: false,
-          title: "More Secure Base Images",
-          description:
-            "Discover best practices for using Chainguard's secure base images in your containerized applications.",
-        },
-        {
-          slug: "chainguard-containers-to-the-rescue",
-          isPath: false,
-          isCourse: true,
-          icon: "bookmark",
-          hasBadge: false,
-          title: "Chainguard Containers To the Rescue!",
-          description:
-            "Explore how Chainguard Containers can help you mitigate vulnerabilities and enhance your security posture.",
-        },
-        {
-          slug: "painless-vulnerability-management-final-exercise",
-          isPath: false,
-          isCourse: true,
-          icon: "bookmark",
-          hasBadge: false,
-          title: "Painless Vulnerability Management: Final Exercise",
-          description:
-            "Apply what you've learned in this comprehensive exercise on vulnerability management.",
-        },
-      ],
-    },
-  ],
-  partners: [
-    {
-      eyebrow: "Partner Training",
-      title: "Chainguard Partner Sales Training",
-      description:
-        "Equip your sales team with the knowledge they need to effectively sell Chainguard's products and solutions.",
-      links: [
-        {
-          isPath: true,
-          isCourse: false,
-          hasBadge: true,
-          title: "Chainguard Discovery: Partner Sales Foundations",
-          slug: "path/chainguard-discovery-partner-sales-foundations",
-          icon: "burger",
-          description:
-            "A comprehensive learning path designed to provide partners with the foundational knowledge needed to effectively sell Chainguard's products and solutions.",
-        },
-        {
-          internalOnly: true,
-          isPath: true,
-          isCourse: false,
-          hasBadge: false,
-          title: "Chainguard Advanced: Partner Sales Accelerator",
-          slug: "path/chainguard-advanced-partner-sales-accelerator",
-          icon: "burger",
-          description:
-            "An advanced learning path aimed at equipping partners with the skills and knowledge to accelerate their sales efforts for Chainguard's products and solutions.",
-        },
-      ],
-    },
-  ],
-  "chainguard-containers-onboarding-guide": [
-    {
-      eyebrow: "Admin Onboarding",
-      title: "Platform Admin Kickoff",
-      description:
-        "Start here to get oriented: kickoff, Console basics, our shared responsibility model, and how to work with Support.",
-      links: [
-        {
-          isPath: false,
-          isCourse: true,
-          hasBadge: false,
-          title:
-            "Kickoff Guide to Chainguard: Getting Started with Our Platform",
-          slug: "kickoff-guide-to-chainguard",
-          description:
-            "A fast, guided kickoff to help you hit the ground running with Chainguard.",
-          icon: "bookmark",
-        },
-        {
-          isPath: false,
-          isCourse: true,
-          hasBadge: false,
-          title: "Getting Started with Chainguard's Console",
-          slug: "getting-started-with-chainguards-console",
-          description:
-            "Learn how to explore, manage, and provision container images through Chainguard’s Console.",
-          icon: "bookmark",
-        },
-        {
-          isPath: false,
-          isCourse: true,
-          hasBadge: false,
-          title: "An Introduction to Chainguard's Shared Responsibility Model",
-          slug: "shared-responsibility-model",
-          description:
-            "See how Chainguard divides security responsibilities so you can focus on building applications while we handle the undifferentiated heavy lifting.",
-          icon: "bookmark",
-        },
-        {
-          isPath: false,
-          isCourse: true,
-          hasBadge: false,
-          title: "Chainguard's Superstar Support",
-          slug: "chainguards-superstar-support",
-          description:
-            "Skip the stress and get answers fast. Learn how to navigate Chainguard’s support like a superstar and keep your supply chain secure with zero ticket anxiety.",
-          icon: "bookmark",
-        },
-      ],
-    },
-
-    {
-      eyebrow: "Technical Onboarding",
-      title: "Secure Image Setup & Ops",
-      description:
-        "Hands-on setup and operations: secure image basics, SSO/IdP integration, registry mirroring, update strategy, EOL planning, customization, migrations, and debugging.",
-      links: [
-        {
-          isPath: false,
-          isCourse: true,
-          hasBadge: false,
-          title: "Chainguard 101: Secure Container Images Basics",
-          slug: "chainguard-101",
-          description:
-            "Learn how to build, run, and maintain apps on secure, minimal Chainguard Containers with confidence.",
-          icon: "bookmark",
-        },
-        {
-          isPath: false,
-          isCourse: true,
-          hasBadge: false,
-          title: "Integrating SSO and IdPs with Chainguard Registry",
-          slug: "integrating-sso-and-idps-with-chainguard-registry",
-          description:
-            "Connect your identity provider and streamline secure access to Chainguard’s registry.",
-          icon: "bookmark",
-        },
-        {
-          isPath: false,
-          isCourse: true,
-          hasBadge: false,
-          title: "Registry Mirroring with Chainguard",
-          slug: "registry-mirroring",
-          description:
-            "Mirror Chainguard images and packages into your internal registry to simplify access, improve reliability, and strengthen control.",
-          icon: "bookmark",
-        },
-        {
-          isPath: false,
-          isCourse: true,
-          hasBadge: false,
-          title:
-            "Keeping Up With Latest: Update Strategies for Chainguard Images",
-          slug: "keeping-up-with-latest",
-          description:
-            "Stay secure and compliant by keeping your Chainguard images up to date—without the chaos.",
-          icon: "bookmark",
-        },
-        {
-          isPath: false,
-          isCourse: true,
-          hasBadge: false,
-          title: "Managing End-of-Life Grace Periods with Chainguard",
-          slug: "managing-end-of-life-grace-periods-with-chainguard",
-          description:
-            "Navigate Chainguard’s EOL Grace Period to keep workloads secure while you plan upgrades.",
-          icon: "bookmark",
-        },
-        {
-          isPath: false,
-          isCourse: true,
-          hasBadge: false,
-          title: "Getting Started with Chainguard's Custom Assembly",
-          slug: "getting-started-with-chainguards-custom-assembly",
-          description:
-            "Quickly and securely customize your Chainguard Images—no Dockerfile required.",
-          icon: "bookmark",
-        },
-        {
-          isPath: false,
-          isCourse: true,
-          hasBadge: false,
-          title: "Migrating to Chainguard Application Images",
-          slug: "migrating-to-chainguard-application-images",
-          description:
-            "Replace vulnerable upstream images with secure, drop-in Chainguard application images in just a few steps.",
-          icon: "bookmark",
-        },
-        {
-          isPath: false,
-          isCourse: true,
-          hasBadge: false,
-          title: "Mastering Base Image Migrations With Chainguard",
-          slug: "mastering-base-image-migrations-with-chainguard",
-          description:
-            "Migrate real apps from “it works on Debian” to minimal, secure Chainguard base images.",
-          icon: ``,
-        },
-        {
-          isPath: false,
-          isCourse: true,
-          hasBadge: false,
-          title: "Containers 102: Practical Image Migration Skills",
-          slug: "containers-102",
-          description:
-            "Hands-on tactics to keep images small, secure, and portable while you migrate to Chainguard.",
-          icon: "bookmark",
-        },
-        {
-          isPath: false,
-          isCourse: true,
-          hasBadge: false,
-          title: "Debugging Chainguard Containers",
-          slug: "debugging-chainguard-containers",
-          description:
-            "Practical strategies to debug Chainguard’s minimal, secure container images in Docker and Kubernetes—without breaking their security model.",
-          icon: "bookmark",
-        },
-      ],
-    },
-  ],
-
-  // Partner intro
-  "chainguard-discovery-partner-sales-foundations": [
-    {
-      eyebrow: "Vulnerability Management",
-      title: "Chainguard Value Proposition",
-      description:
-        "Lay the foundations for your understanding of how to best position Chainguard in the open source space with consideration for vulnerability management pain.",
-      links: [
-        {
-          isPath: false,
-          isCourse: true,
-          hasBadge: false,
-          title: "Software Vulnerabilities: What Are They?",
-          slug: "software-vulnerabilities-what-are-they",
-          description:
-            "Understand the basics of software vulnerabilities, including common types and their implications.",
-          icon: "bookmark",
-        },
-        {
-          isPath: false,
-          isCourse: true,
-          hasBadge: false,
-          title: "How to Manage CVEs",
-          slug: "how-to-manage-cves",
-          description:
-            "Learn best practices for managing Common Vulnerabilities and Exposures (CVEs) in your software supply chain.",
-          icon: "bookmark",
-        },
-        {
-          isPath: false,
-          isCourse: true,
-          hasBadge: false,
-          title: "Chainguard Containers Crash Course",
-          slug: "linkys-crash-course-on-chainguard-images",
-          description:
-            "A comprehensive introduction to Chainguard Containers and their benefits.",
-          icon: "bookmark",
-        },
-      ],
-    },
-    {
-      eyebrow: "Pitching Chainguard",
-      title: "How to Effectively Pitch Chainguard",
-      description:
-        "Learn strategies and best practices for effectively pitching Chainguard to potential customers.",
-      links: [
-        {
-          isPath: false,
-          isCourse: true,
-          hasBadge: false,
-          title: "Pitching the Chainguard Way",
-          slug: "pitching-the-chainguard-way",
-          description:
-            "A walk-through of key points for pitching Chainguard solutions.",
-          icon: "bookmark",
-        },
-      ],
-    },
-    {
-      eyebrow: "Selling Chainguard",
-      title: "Quoting Chainguard Products",
-      description:
-        "Learn how to effectively quote Chainguard products and solutions to customers.",
-      links: [
-        {
-          isPath: false,
-          isCourse: true,
-          hasBadge: false,
-          title: "Partner Guide to Chainguard Pricing",
-          slug: "partner-guide-to-chainguard-pricing",
-          description:
-            "Understand the basics of quoting Chainguard products and solutions to customers.",
-          icon: "bookmark",
-        },
-      ],
-    },
-    {
-      eyebrow: "Show off your skills",
-      title: "Claim your badge",
-      description:
-        "After completing this learning path, don't forget to claim your Chainguard Partner Sales Foundations badge to showcase your expertise.",
-      links: [
-        {
-          isPath: false,
-          isCourse: true,
-          hasBadge: true,
-          title: "Chainguard Discovery: Partner Sales Feedback and Support",
-          slug: "chainguard-discovery-partner-sales-feedback-and-support",
-          description:
-            "Provide feedback, find support resources, and claim your badge after completing the Chainguard Partner Sales Foundations learning path.",
-          icon: "bookmark",
-        },
-      ],
-    },
-  ],
-};
+export { footerData } from "../data/footer.mjs";
+export { pathSections } from "../data/path-sections.mjs";
